+++ bortus-manager/src/main/java/com/orvillex/bortus/manager/modules/scheduler/domain/JobLogReport.java
@@ -24,7 +24,7 @@ public class JobLogReport {
 
     private Date triggerDay;
 
-    private int runningCount;
-    private int sucCount;
-    private int failCount;
+    private Long runningCount;
+    private Long sucCount;
+    private Long failCount;
 }
